backend/requirements.txt
Django==5.0.6
pytz>=2024.1
sqlparse>=0.5.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
# JWT auth
djangorestframework-simplejwt==5.3.1
